requirements.txt
pandas==2.2.2
matplotlib==3.9.0 
transformers-lightning==0.7.12
kaggle==1.6.14
pyzipper==0.3.6
ipykernel==6.29.4
seaborn==0.13.2
huggingface-hub==0.23.4
datasets==2.20.0
beautifulsoup4==4.12.3
Unidecode==1.3.8
flashtext==2.7
emot==3.1
sentence-transformers==3.0.1
swifter==1.4.0
scikit-learn==1.5.0
umap-learn==0.5.6
plotly==5.22.0
nbformat==5.10.4
pytorch-lightning==2.3.0
memory_profiler==0.61.0
torch-tb-profiler==0.4.3
protobuf==4.25
torcheval==0.0.3
torchcp==0.1.0
tensorboardX==2.6.2.2
MAPIE==0.8.6
openpyxl==3.1.5
